separate tests requireing GRASS from the others
