+++ .github/workflows/build_wheels.yml
@@ -28,13 +28,13 @@ jobs:
         env:
           CIBW_ARCHS: auto64  # Don't build 32 bits wheels
           CIBW_SKIP: pp*  # Don't build pypy wheels
-          CIBW_ENVIRONMENT: "ITZI_BDIST_WHEEL=1"
+          CIBW_ENVIRONMENT: "ITZI_BDIST_WHEEL=1"  # Trigger generic compiler flags
           CIBW_ENVIRONMENT_MACOS: >
             MACOSX_DEPLOYMENT_TARGET=14.0
             ITZI_BDIST_WHEEL=1
           CIBW_BEFORE_TEST: "pip install pytest pandas"
           CIBW_TEST_SOURCES: tests
-          CIBW_TEST_COMMAND: >
+          CIBW_TEST_COMMAND: >  # Fast tests which do not require GRASS
             pytest {package}/tests/test_flow.py {package}/tests/test_analytic.py {package}/tests/test_rastermetrics.py
 
       - uses: actions/upload-artifact@v4
